+++ src/compatibility.ts
@@ -4,6 +4,7 @@ export const SUPPORTED_DSH_VERSIONS = Object.freeze([
   '0.1.0-rc.6',
   '0.1.0-rc.7',
   '0.1.1-rc.2',
+  '0.1.2-alpha.1',
 ] as const)
 
 /**